+++ ament_archlint/ament_archlint/pytest_marker.py
@@ -1,5 +1,4 @@
-# Copyright 2025 PAL Robotics S.L.
-# Copyright 2015 Open Source Robotics Foundation, Inc.
+# Copyright 2025 PAL Robotics
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,8 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-file(GLOB_RECURSE _source_files FOLLOW_SYMLINKS "package.xml")
-if(_source_files)
-  message(STATUS "Added test 'skilllint' to check PAL module manifests")
-  ament_skilllint()
-endif()
+
+def pytest_configure(config):
+    config.addinivalue_line(
+        'markers',
+        'archlint: marks tests checking manifests of skills being conformant')