+++ ament_archlint/CHANGELOG.rst
@@ -1,5 +1,5 @@
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Changelog for package ament_skilllint
+Changelog for package ament_archlint
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 0.3.1 (2025-06-18)
@@ -15,5 +15,5 @@ Changelog for package ament_skilllint
 0.2.0 (2025-01-15)
 ------------------
 * add unittests for validate_manifest
-* create ament_skilllint and ament_cmake_skilllint to integrate skill manifest checking to ament test
+* create ament_archlint and ament_cmake_archlint to integrate skill manifest checking to ament test
 * Contributors: Séverin Lemaignan